requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
django-cors-headers==4.8.0
python-slugify==8.0.4
Pillow==11.3.0
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
django-filter==24.3